// UCManual1/src/webparts/tiles/components/Tiles.module.scss.d.ts
declare const styles: {
  tiles: string;
  teams: string;
  tilesContainer: string;
  tile: string;
  tileImageContainer: string;
  tileImage: string;
  tileContent: string;
  tileTitle: string;
  tileDescription: string;
  emptyTile: string;
  placeholderContent: string;
  placeholderIcon: string;
  placeholderTitle: string;
  placeholderText: string;
};

export default styles;